streamlit/requirements.txt
functions_framework
Flask
pandas
numpy
google-cloud-bigquery
google-cloud-storage
google-auth
google-auth-oauthlib
streamlit
sqlparse
langfuse
openai
vacances_scolaires_france==0.10.0
plotly==5.23.0
prophet
scipy==1.14.0
scikit-learn
plotly
google-cloud-documentai
pdf2image
opencv-python
matplotlib
watchdog